requirements.txt
anyio==4.5.2
autopep8==1.6.0
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.90.1
flake8==6.1.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
mccabe==0.7.0
passlib==1.7.4
pycodestyle==2.11.1
pydantic==1.10.7
pyflakes==3.1.0
sniffio==1.3.1
SQLAlchemy==2.0.0
starlette==0.23.1
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.18.3
python-jose==3.3.0
python-dotenv==0.21.0
cryptography==39.0.1
alembic==1.11.0
python-multipart==0.0.5